Docs: Begin refactoring of developer protocol documentation.
This is the beginning of a complete rewrite of the "Multiplayer Packet
Structure" page, with more Obsidian-ism to make traversing nice.

This is only the first refactor so hopefully we'll be able to come up
with a common scheme and layout soon.

Something I just noticed that I have missed is the direction in which
these packets are sent (Client -> Server, Server -> Client, or Server
<-> Client)